tcl/target/icepick: remove unnecessary ir call
The icepick_c_router call includes a 6-byte write to the IR of a value
of 2. The document "Router_Scan_Sequence-ICEpick-D.pdf" clearly
indicates this is necessary to send before the first shift DR, but the
process does not include another write before the 0xeQ002008 register.

Remove the unnecessaly IR Scan.